SMW-341 Update ut-alert folder structure  * Add joi validations for alert.queueOut.read
