+++ api/sql/schema/381$alterScriptMessageIn.sql
