+++ api/sql/schema/750$alert.template.fetch.sql
