+++ validations/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+    'queueOut.read': require('./queueOut/read')
+};
